[r] Move identical integration test into a separated file
